requirements.txt
streamlit==1.38.0
pandas==2.2.3
google-generativeai==0.7.2
kaggle==1.6.17
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
google-genai==0.3.0
